Migrated fetch api call from Search context to individual search components
